#32 (front/refacto) update todolist, extract todo creation, moved into new createTodo component, both are inside root component (App)
