+++ frontend/src/components/Todolist.vue
@@ -1,47 +1,19 @@
 <template>
-  <div class="container" v-if="isLoaded" style="width: 450px">
-    <v-list class="elevation-2">
-      <v-list-subheader color="blue">TODOLIST</v-list-subheader>
-      <transition-group v-if="todos.length > 0" name="list" tag="div" >
+  <v-list class="elevation-2" v-if="isLoaded">
+    <v-list-subheader color="blue">TODOLIST</v-list-subheader>
+    <transition-group v-if="todos.length > 0" name="list" tag="div">
       <todo-item
           v-for="(todo) in todos"
           :key="todo.id"
           :id="todo.id"
       />
-      </transition-group>
-      <v-list-item v-else>
-        <template v-slot:title>
-          Aucune tâche à réaliser
-        </template>
-      </v-list-item>
-    </v-list>
-    <v-btn
-        color="blue"
-        @click="dialogAddTodo = true"
-    >
-      Créer une nouvelle tâche
-    </v-btn>
-    <v-dialog
-        v-model="dialogAddTodo"
-        width="300px">
-      <v-card>
-        <todo-input
-            :autofocus="true"
-            @validate="addTodo">
-        </todo-input>
-        <v-card-actions>
-          <v-spacer></v-spacer>
-          <v-btn
-              color="blue-darken-1"
-              variant="text"
-              @click="dialogAddTodo = false"
-          >
-            Fermer
-          </v-btn>
-        </v-card-actions>
-      </v-card>
-    </v-dialog>
-  </div>
+    </transition-group>
+    <v-list-item v-else>
+      <template v-slot:title>
+        Aucune tâche à réaliser
+      </template>
+    </v-list-item>
+  </v-list>
 </template>
 
 <script lang="ts">
@@ -55,11 +27,9 @@ export default {
 import {onBeforeMount, ref} from "vue";
 import {useTodoStore} from "@/stores/todo";
 import {storeToRefs} from "pinia";
-import TodoInput from "@/components/TodoInput.vue";
 import TodoItem from "@/components/Todo.vue";
 
 const isLoaded = ref(false);
-const dialogAddTodo = ref(false);
 const store = useTodoStore();
 const {todos} = storeToRefs(store);
 
@@ -67,20 +37,9 @@ onBeforeMount(async () => {
   await store.loadTodos();
   isLoaded.value = true;
 })
-
-async function addTodo(description: string) {
-  await store.addTodo(description);
-  dialogAddTodo.value = false;
-}
 </script>
 
 <style scoped>
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: 25px;
-  margin-top: 100px;
-}
 
 .list-enter-active,
 .list-leave-active {
